First version of Parallel Coherent Hash. Debug mode. Warnings corrected.

+++ include/timer.h
@@ -1,13 +1,13 @@
+/* -*- c++ -*- */
 #pragma once
 
 #include <ctime>
-
 #include "omp.h"
 
 double GetCurrentTime() {
 #ifndef _OPENMP
-	return std::clock() / static_cast<double>(CLOCKS_PER_SEC);
+  return static_cast< double >( std::clock()) / static_cast< double >( CLOCKS_PER_SEC );
 #else
-	return omp_get_wtime();
+  return omp_get_wtime();
 #endif
 }
